すべてをコンポーネントに移動させる
